support float in number inputs
Integer can still be parsed and are converted back to the same string
without decimal places in JavaScript, e.g.

    String(Number.parseFloat("1"))

returns "1" and not "1.0".

+++ robot_cameraman/server_static_folder/limits-menu.js
@@ -185,8 +185,8 @@ class LimitsMenu extends HTMLElement {
           [configurationKey]:
             activeElement.checked
               ? [
-                Number.parseInt(minimumRangeElement.value),
-                Number.parseInt(maximumRangeElement.value),
+                Number.parseFloat(minimumRangeElement.value),
+                Number.parseFloat(maximumRangeElement.value),
               ]
               : null,
         }
